Move redundant unit test helper code to tools
The classes TempZimArchive and TestItem needs to be reused efficiently.
Moving it into tools will help us to take advantage of it without
duplicacy.
