+++ test/search.cpp
@@ -22,9 +22,6 @@
 #include <zim/archive.h>
 #include <zim/item.h>
 #include <zim/search.h>
-#include <zim/writer/creator.h>
-#include <zim/writer/item.h>
-#include <zim/writer/contentProvider.h>
 
 #include "tools.h"
 
@@ -35,37 +32,8 @@ namespace
 {
 
 using zim::unittests::getDataFilePath;
-
-class TestItem : public zim::writer::Item
-{
-  public:
-    TestItem(const std::string& path, const std::string& title, const std::string& content):
-     path(path), title(title), content(content)  { }
-    virtual ~TestItem() = default;
-
-    virtual std::string getPath() const { return path; };
-    virtual std::string getTitle() const { return title; };
-    virtual std::string getMimeType() const { return "text/html"; };
-
-    virtual std::unique_ptr<zim::writer::ContentProvider> getContentProvider() const {
-      return std::unique_ptr<zim::writer::ContentProvider>(new zim::writer::StringProvider(content));
-    }
-
-  std::string path;
-  std::string title;
-  std::string content;
-};
-
-// Helper class to create a temporary zim and remove it once the test is done
-class TempZimArchive : zim::unittests::TempFile
-{
-  public:
-    explicit TempZimArchive(const char* tempPath) : zim::unittests::TempFile {tempPath} {}
-
-    const std::string getPath() {
-      return this->path();
-    }
-};
+using zim::unittests::TempZimArchive;
+using zim::unittests::TestItem;
 
 #if WITH_TEST_DATA
 TEST(Search, searchByTitle)
@@ -93,7 +61,7 @@ TEST(Search, indexFullPath)
   creator.configIndexing(true, "en");
   creator.startZimCreation(tza.getPath());
 
-  auto item = std::make_shared<TestItem>("testPath", "Test Article", "This is a test article");
+  auto item = std::make_shared<TestItem>("testPath", "text/html", "Test Article", "This is a test article");
   creator.addItem(item);
 
   creator.setMainPath("testPath");
